Update patch set 2
Patch Set 2:

(1 comment)

> Patch Set 2:
> 
> Would it be easier to just do it as part of the make file (without an additional dex2oat arg)?
> (like here: https://android-review.googlesource.com/c/platform/art/+/854070)

No. The makefiles/soong has a lot of bs expansions for options and other things and I don't want to have to figure out how to get everything expanded.

> Or you see this useful outside the makefiles?

not really. This is mostly for personal testing of hermetic builds. I was thinking of using it to make tests but that's still a little ways away.

Patch-set: 2
Reviewer: Gerrit User 1047769 <1047769@85c56323-6fa9-3386-8a01-6480fb634889>