Update patch set 2
Patch Set 2:

(2 comments)

> Patch Set 1:
> 
> (1 comment)

Yes, this style is consistent in this file.

The goal seems fine here. We'll need to rebase the change against aosp.

We are moving to a new process for upstreaming requests (e.g. require a partner buganizer entry to track) so we don't lose track of change requests.

Patch-set: 2
CC: Gerrit User 1071150 <1071150@85c56323-6fa9-3386-8a01-6480fb634889>