Update patch set 10
Patch Set 10:

> Patch Set 10:
> 
> It looks like you decided to handle this completely differently from TargetHeapUtilization?
> 
> Is there a reason for that? I think they should be logically closely connected: If you set the evacuation threshold too low, TargetHeapUtilization may not be achievable.
> 
> (In fact, we don't seem to pay as much attention to TargetHeapUtilization as one might assume, so this is only somewhat true.)

The main motivation to make the live percent evacuation threshold a run-time parameter is so that we can test various values before choosing one that is satisfying.

We should probably make TargetHeapUtilization a run-time parameter as well. Do you think they should be coupled? Or that we should introduce constraints on a pair of acceptable values for both settings?

Patch-set: 10
Reviewer: Gerrit User 1052644 <1052644@85c56323-6fa9-3386-8a01-6480fb634889>