WIP: RFC: ART: Introduce non-precise stackmap.
Some of safepoints don't need to have precise vreg info;
this will decrease the stackmap size.

.oat file size reduction:
 - boot.oat:           -244 kb (-5.8%)
 - boot-framework.oat: -736 kb (-5.1%)

DO NOT MERGE
The intention of this patch is to open the discussion and show the
impact, not to be a final solution.

One of the possible solutions could be to introduce is_precise
to HEnvironment and properly support it in optimization passes and
code generator; CodeInfo class might need some changes as well.

Test: test-art-target, +gc-stress.
Change-Id: Idbad355770e30a30dcf14127642e03ee666878b8