Auto Fast JNI Detection for x86 & x86_64
This patch implements Binary Analysis for x86 during native method registration at runtime.
Marking methods as fast means the thread does not have to do costly transition from kRunnable to kNative.
This will reduce the overhead of JNI transition from Java to Native
The method is classified as fast if it satisfies certain constraints which we have mentioned as below .As of now we have checked as these
minimal gating conditions to start of :
 - No indirect calls or jumps
 - No more than 3 levels of call nesting
 - No unsupported instructions,or interrupts
 - No repeats(or loops),
 - No instructions with LOCK prefix
 - Less than or equal to 3 levels of call depth nesting (20 Basic Blocks & 100 Instructions)
Essentially, all simple Native methods that satisfy the above criteria are marked as fast.
Disassembly of x86 Instructions is done using the Capstone Library.( for this we have added a separate patch)
The auto fast JNI is enabled by only in JIT mode and currently does the analysis for x86 only instruction set
We have observed performance runtime benefits:
=> jArtBand JNI  benchmark - approx 30%
Jave Apps ( meeting the gating condition )Launch time for Recursive Launches( second or third launch) :
-> Twitter , UC browser : approx 5-6%
Test: ./test.py --host --64

Change-Id: Iae34724c31af1cf836b5dfa5e3be273a820d9e5a
Signed-off-by: Priyanka Bose <priyanka.bose@intel.com>