Implementation of partial loop unrolling with unknown iterations for x86_64.
Implementation for other instruction-sets is kept as it is.

Current implementation supports loops having 2 blocks.
This patch is improving below benchmarks scores with Intel Atom chipset
    CaffeineMark.Loop by 6%, CaffeineMark;Overall by 1%,
    SPECjvm2008-1T;scimark.sparse.large by 5%,
    SPECjvm2008-4T;scimark.sparse.large by 12%,
    SPECjvm2008-4T;scimark.lu.small by 4%,
    SPECjvm2008-4T;compiler.compiler by 5%.

With patch, we observed .3% increment in size of boot.oat and .1% increment in
size of boot-framework.oat files.

With patch, we observed ~1% increment in compile time of boot.oat file.

Test: ./test.py --host --64

Change-Id: I9d8eb4e49b70330fec55fc7781958faf68ed713e
Signed-off-by: Neeraj Solanki <neeraj.solanki@intel.com>