Update patch set 6
Patch Set 6:

> Patch Set 6:
> 
> These are interesting suggestions, but different to the upstream request made.
> 
> Initial thoughts:
> 
> 1) under load may be the timeout may not be ideal so this could introduce new failure modes which we'd need to consider. ART has little specialization for system server.
> 
> 2) there is no requirement to use pre-built runtime module during device bring up. This kind of change seems like a temporary debugging aid for a driver or kernel bug.
> 
> My preference would be to land this CL as an incremental improvement and then look at the new proposal in a separate upstream request (should be made with TAM using upstreaming guidelines).


1.
we only hope not FATAL here in debug build , ERROR is enough 
like this:

-          LOG(kIsDebugBuild ? ::android::base::FATAL : ::android::base::ERROR)
+          LOG(::android::base::ERROR)


2.
There are often some modifications and optimizations in our kernel code.
So we occasionally encounter such problems.But such problems occur less frequently.

If only print ERROR here, the process still in the phone ,we can continue debug.

If FATAL here, 
For art is going to close the source and give lib only ,we have to make a debug lib and reTest it.
But very low frequency of occurrence, will greatly reduce the efficiency of analysis and resolution of such problems.

3.
We still insist on suggestions not FATAL here, use ERROR is better.
This will bring us great help.

Patch-set: 6
Reviewer: Gerrit User 1130937 <1130937@85c56323-6fa9-3386-8a01-6480fb634889>