Update patch set 2
Patch Set 2:

It sounds to me like we need some additional mutual exclusion to avoid this scenario. Concurrent modification of the result of GetMethods should be impossible. Mathieu - Do you understand whether we still have this problem, and whether there is a way to avoid this scenario?

I'm not a fan of this solution, since it doesn't avoid the underlying race; it merely reduces the probability of it resulting in a crash; the code is still incorrect. If the concurrent update happens just after the num_methods check, we still have the same problem.

Patch-set: 2
Reviewer: Gerrit User 1042828 <1042828@85c56323-6fa9-3386-8a01-6480fb634889>