Revert "Bionic libs are located at /system/lib"
This reverts commit 3f20fe74da1df4391f5c70f7124793068e261544.

Reason for revert: Breaks libcore tests: b/122513239. The problem is that
binaries in the Runtime APEX will get duplicate instances of libc/dl/m as
long as the libraries are still present in /apex/com.android.runtime/${LIB},
so http://r.android.com/860445 can only go in when they are gone.

Bug: 122513239
Bug: 120266448
Test: mmm external/vogar && vogar --verbose libcore/luni/src/test/java/libcore/libcore/icu/TimeZoneIntegrationTest.java
Test: atest --test-mapping cts/tools/vm-tests-tf/TEST_MAPPING
Test: atest CtsJniTestCases CtsCompilationTestCases
Change-Id: Ia535596cc3ec7dee98c03ed6fcfb23169d8f3e3e