Update patch set 1
Patch Set 1:

This is the first step towards using the Runtime APEX in ART's standalone test setup. I'll also make the required changes on the ART Buildbot to use the new scripts, and adjust the test recipes to use the artifacts from the Runtime APEX (instead of the ones in `$ART_TEST_CHROOT/system`).

Ideally, it'd be nice to be able to use apexd (or maybe a wrapper around libapexd) to do the package activation/deactivation work (instead of loosely emulating it). We can think about it later.

Patch-set: 1
Reviewer: Gerrit User 1052644 <1052644@85c56323-6fa9-3386-8a01-6480fb634889>