Honor `ART_TEST_ANDROID_RUNTIME_ROOT` when running ART gtests on device.
When this environment variable is defined, use its value to define
`ANDROID_RUNTIME_ROOT` when running ART gtests on device.

This is a generic change, but its first intended use is to point
`ANDROID_RUNTIME_ROOT` to "/system" when running gtests on target, so
that ICU can find ICU .dat files in "/system/etc/icu", until the
Runtime APEX is available in target testing on the ART Buildbot.

Test: Rely on the ART Buildbot.
Bug: 121117762
Bug: 119293618
Bug: 113373927
Change-Id: I96e12e237d3c722609a042aaf1a946b654f9544f