Update patch set 3
Patch Set 3:

Hi!

First of all I still feel very glad that the SuperblockCloner and loop optimizations are being considered and investigated for further optimizations; I would be more than happy to answer all the questions and comments on the original contribution and possible improvements!

In regard to this particular CL could you please describe what particular semantic problem you are trying to solve (which can't be solved by the current implementation) and how? Tests and examples would be very valuable to achieve that and to explain your approach!

Patch-set: 3