Update patch set 7
Patch Set 7:

> Patch Set 5:
> 
> What is the impact of making systrace tags significantly more expensive with this indirection?

We could change the API to have a set up method to get the address of the flags and the mask value? This won't claw back everything, but might be better.

That's purely speculative as this seems to fail in go/lem for now.

Patch-set: 7
Reviewer: Gerrit User 1071150 <1071150@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0