Update patch set 9
Patch Set 9:

As discussed, I'm wondering we drop this CL entirely, it's just a nonsense that was conceived before we had a good handle on the (trivial) dependencies. Maybe we persevere with this short term, but aim to deprecate asap.

Alternatives might be growing the one of the ndk flavors or building mini-stubs for the functions/libraries used.

Patch-set: 9
Reviewer: Gerrit User 1071150 <1071150@85c56323-6fa9-3386-8a01-6480fb634889>