Update patch set 12
Patch Set 12:

(9 comments)

Sorry for the late review, I'm fine to descope all of these comments if you're anxious to get it in.

> Patch Set 9:
> 
> As discussed, I'm wondering we drop this CL entirely,

We still need the C APIs, which is the meat of this CL, right?

> it's just a nonsense that was conceived before we had a good handle on the (trivial) dependencies. Maybe we persevere with this short term, but aim to deprecate asap.

Yeah, or perhaps just shave off the dlopen stuff in favour of APEX stubs.

Patch-set: 12
Reviewer: Gerrit User 1373864 <1373864@85c56323-6fa9-3386-8a01-6480fb634889>