Ensure the correctness of fast verify using sdk-verification
We cannot guarantee that the resolution recorded in the vdex file is
correct when the boot classes are redefined. For example we might be doing
only sdk verification and not have the entire context available - as such,
if an app redefines a class that is in the boot classpath but not the sdk
we might record the wrong resolution.

The CL discards the vdex file when the app redefines classes from boot
classpath. Currently this is done only when the vdex file comes from a
DexMetadata archive.

Test: test/testrunner/testrunner.py -t 719
Bug: 122968669
Change-Id: If0c56b1970d8ebe701d198ffccec52f586aea9e6