Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> What would this buy us? Do you have a specific scenario in mind?

Scenario: App is upgraded, changes targetSdkVersion. That changes the SELinux label of "oat/" and as a result, installd does a recursive restorecon() on the whole directory. The oat file is still up-to-date wrt the secondary dex but now has a non-exec label and cannot be loaded executable.

Patch-set: 1
Reviewer: Gerrit User 1059815 <1059815@85c56323-6fa9-3386-8a01-6480fb634889>