Open libdexfile_external on demand in libdexfile_support.
This allows users to depend on libdexfile_support (typically via
libbacktrace) also in binaries that don't run java code. The drawback is one
extra address load in all calls.

Also add support for a nonfunctional fake with recovery and vendor variants,
so that clients don't need special cases to avoid the libdexfile dependency.

Bug: 123403798
Bug: 123186083
Test: Flash and boot
Test: atest -a libdexfile_support{,_static}_tests
Test: mmma art/libdexfile/
Test: mmma system/core/{libunwindstack,libbacktrace}, run host gtests (cannot get atest to work)
Change-Id: Ia99ad8039bb37fe25bf7009a85ae89c2aa660bd1