Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Not happy with this. This asserts that a *valid* change that breaks an agent should block JVMTI, putting the burden of maintenance on the ART team instead of the owner of the agent.

If mockito breaks, a significant section of AOSP testing and a some 3rd party testing will break. Hence we should very strongly avoid having mockito in a broken state.

If the Agent is buggy or a test is flaky, this is the agents fault. Hence every time we update the agent or test, we run these tests. Until now these tests were 100% reliable (either broken or success) and never flaky.

If JVMTI is changed there is a small likelyhood that the API changed. If this is an unintentional change and this breaks the agent - good that we found it. If this is an intentional change, we have the rare case of an intentional API change. In this case the agent has to be changed too. As describe above we cannot leave mockito broken, hence in this case we have to change the agent in lock step with the API change.

I think the contentious part is the agent abusing a jvmti API and a valid change to JVMTI change cannot be submitted as mockito would be broken because of this badly implemented agent. In this case the agent need to be fixed. We can fix the agent before we submit the JVMTI change, delaying JVMTI work. We can submit the JVMTI change first and then fix the agent. The problem with the second option is that a large part of AOSP testing will be broken. Yes it will not be your fault, but we cannot just break everybodies tests.

Patch-set: 1
Reviewer: Gerrit User 1084818 <1084818@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0