Generalize vdex class redefinition check
The check introduced in CL If0c56b1970d8ebe701d198ffccec52f586aea9e6
skips fast verification if an apk's class is overshadowed by a class in
boot classpath because the vdex dependencies do not contain intra-apk
dependencies.

However, the change only checks for presence of a duplicate class in the
boot classloader, while a duplicate class could be in any of the parent
classloaders. Fix this and move the check into VerifierDeps to make it
a proper part of the verification process.

The CL also refactors VerifierDeps::ValidateDependencies to output
an error string for better logging.

Bug: 122968669
Test: test/testrunner/testrunner.py -t 719
Change-Id: I0d06b82e31088c58d4493723a5435309740f1d0c