WIP: Add stand-alone tool to generate mini-debug-info.
Reuse ART's infrastructure to create stand-alone command line
tool which can be used to create and optimize mini-debug-info
for native libraries as part of the build process.

Create small static library to hold most of our elf writing
and reading classes.  This avoids depending on the runtime.

Remove MIPS support from the ELF builder.  This is slightly
easier than making it independent of the runtime.

Bug: 110133331
Change-Id: I93343808d0e27ee8e1117e713a2503e8179fc245