Fix vdex fast-verify performance regression
Recent CL I0d06b82e31088c58d4493723a5435309740f1d0c generalized the
fast-verify class redefinition check by checking that all vdex-verified
classes resolve to dex files covered by the vdex and are not duplicates
of classes in parent class loaders. This introduced a performance and
allocated memory regression for dex2oat invoked with
compiler-filter=verify(-profile) because mirror::Class objects were not
being created for all app classes.

This patch removed the memory size regression and reduced the
performance impact by refactoring ClassLinker::FindClass* functions and
introducing ClassLinker::FindClassDexFile() which finds the class but
does not load it.

It further reduces the performance impact by stopping iteration when the
first vdex dex file is encountered, assuming that it is known the class
would be found there. This could otherwise have significant impact on
large apps.

A small performance overhead remains on Golem as the benchmark puts
framework in the same PCL as the compiled app and the original check
would only attempt to resolve in boot class loader.

Test: run `dex2oat filter=interpret-only; dex2oat filter=verify`
      compare time and allocated memory numbers before CL and after
Change-Id: Ifd690cdafdc99d3eafb9847d67775fc11a5b5023