Introduce vdex-based verification caching for InMemoryDexClassLoader
When loading dex files using InMemoryDexClassLoader, verification can be
a bottleneck (up to 50% increase in app startup time observed, compared
with PathClassLoader). Address this issue by caching verification
results in a vdex file in the app's data folder.

On first load, libcore will fully initialize the class loader and call
DexFile.preverify() at the very end. At this point, OatFileManager
spawns a new thread and attempts to verify all classes defined in the
newly loaded dex files and writes the information into a vdex file.
The path of the vdex file is determined from the path to the app's
data folder provided by the framework via VMRuntime at process setup.
The name of the file is the combined adler32 checksum of all loaded
dex files.

Only the bitvector of verified/unverified classes is stored in the vdex,
assuming that the vdex will be fully recreated if it ever becomes
out-of-date. To detect that, the vdex format was also extended with
boot class path checksums and class loader context (only used in this
code path for the moment). This also means that:
(a) dex files are not stored in the vdex
(b) no data about the dex files are leaked other than dex file checksums
and a list of class def inidices which failed verification. As a result,
the files are typically less than 100 bytes in size.

On subsequent loads, ART will detect the presence of the vdex and will
(a) skip structural verification of the dex files,
(b) initialize a subclass of OatFile called OatFileBackedByVdex which
parses the class verification status data from the vdex and initializes
OatClass instances to Status::kVerified / kNotReady accordingly.

There is an upper limit on the number of vdex files in the data folder,
currently set to 8. When this threshold is reached, ART will unlink the
least recently accessed anonymous vdex file (atime returned by stat()).

Bug: 72131483
Test: art/test.py -r -b -t 691 -t 692
Test: art/tools/run-libcore-tests.sh
Change-Id: I1d82dc441c35c7b5db04038eeaf74c3a5dcd64be