Add mean peak RSS for GC metric to GC perf dump
Use libmeminfo to extract /proc/self/smaps info and compute RSS specific
to GC data structures and heaps. It must be computed at a time in the GC
cycle when the physical memory consumption is at its peak. For instance,
for CC it is immediately after marking is finished and before regions
are cleared.

Test: art/test/testrunner/testrunner.py --target --runtime-option=-XX:DumpGCPerformanceOnShutdown
Change-Id: I92cf006524cf6c91ba1e96aa7c5303c578e6db54