Assign conscrypt to core-platform domain, other apex to platform
Until now only /system/framework and /apex/com.android.runtime were
known locations, assigning "platform" domain to the former and
"core-platform" domain to the latter. The media and conscrypt modules
were left in the "application" domain.

This patch adds a known path to the conscrypt module, configurable via
ANDROID_CONSCRYPT_ROOT env variable, and moves it to "core-platform".
The media module is covered by testing is location is in "/apex/".
All such dex files are assigned to "platform" domain.

Any other dex files on boot classpath are also assigned to "platform"
on first access. A warning is printed in such case and it is now deduped
to avoid logspam.

Bug: 125701194
Bug: 119068555
Test: compiles, boots, no conscrypt/media warnings
Test: art/test.py -b -r -t 674
Change-Id: I33bef18459741095d3d99b541fc88b21cf547800