Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Good CL, but I'm not convinced this is desirable.
> 
> Can we not provide a suitable API surface in the intra core API for the relevant classes? Or even make the appropriate surface area non-hidden.

We wanted a mechanism for reporting calls from framework into the managed core libraries to methods that are not marked as core platform API to highlight to OEMs where they (or code they've imported) is straying outside of our stability guarantees. From this perspective, conscrypt is just like anything else in the core managed library and needs to be exempted from triggering it. To be clear, this violation reporting is not intended to enforce _runtime module_ boundary, but the _managed core library_ boundary.

Enforcing / preventing violations between the conscrypt module and the runtime module, i.e. the intra-core API, is another discussion. I'd actually be averse to putting anything in that increased complexity or added a runtime penalty because we are/will be (for the most part) responsible for the content of both conscrypt and the runtime module and any violations are by us on us.

Patch-set: 1
Reviewer: Gerrit User 1038039 <1038039@85c56323-6fa9-3386-8a01-6480fb634889>