hiddenapi domain: Use oat dex file location if available
To determine the caller and callee domains of a hidden API access check,
the logic will test the dex location against several known system
locations. However, DexFile instances backed by an OatFile have relative
dex locations to avoid need to rewrite the .oat file when moving files
between folders. The canonical dex location is stored in the OatDexFile
instead.

Because the OatDexFile data structure is not available in libdexfile,
move the domain resolution logic to hidden_api.cc and resolve lazily on
first access. To this end, the enum of domains is extended with
kUnresolved, a new default value.

Test: m test-art-gtest
Test: camera does not crash
Bug: 126901248
Change-Id: Id494b1c47a2199c227dec046174e08320b9cbc3b