Preload const-strings only for profile methods.
This fixes a boot image size regression due to including all
the const-strings.

Note that this re-introduces one source of non-determinism
when compiling the boot image, i.e. string resolution for
inlined methods not present in the profile. This shall be
fixed by a follow-up CL.

Boot image sizes for aosp_taimen-userdebug:
 - before:
   arm/boot*.art: 14483456
   arm64/boot*.art: 18751488
 - after:
   arm/boot*.art: 12345344 (-2.0MiB)
   arm64/boot*.art: 16609280 (-2.0MiB)

Test: aosp_taimen-userdebug boots.
Bug: 76145463
Bug: 26687569
Change-Id: I5127d58bc7bdf24532fda6cc7d32a1bd2068452e