Update patch set 1
Patch Set 1: Code-Review-2

This patch is likely unneeded, because of a wish by security to keep the domain tied down. A patch that's purely framework (that will load things in the _startup domain before this is set) is incoming. -2 for now.

Patch-set: 1
Reviewer: Gerrit User 1041833 <1041833@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-2
Label: Presubmit-Ready=0