Set apex_name property to com.android.runtime
The property is used to let the build system to know the canonical name
of an APEX which might be different from the module name of it. In case
of the runtime APEX, the module name is
com.android.runtime.{debug|release} whereas the canonical name is
com.android.runtime.

The canonical name should be known to the build system in order for it
to install symbol files to the paths where the files in an APEX will be
activated on. For example, the symbol file for ./lib/libFoo.so in the
runtime APEX should be installed to
$(PRODUCT_OUT)/symbol/apex/com.android.runtime/lib/libFoo.so

Note that apex_name property is NOT set for com.android.runtime.release,
because doing so will create multiple rules from *.release and *.debug
having the same target path, which breaks the build. Since the symbol
files are usually (but not always) required for debug builds, it is set
in the *.debug APEX only.

Bug: 120846816
Test: m and inspect that symbol files exist under
$(PRODUCT_OUT)/symbols/apex/com.android.runtime/

Change-Id: If3969736ed1c51b8375974dfa320cd798dbd730c