Update patch set 2
Patch Set 2:

> Patch Set 2: Code-Review-2
> 
> -2 because we need the check for FSI.

(This requires a longer conversation so gerrit is probably not the right place for this. The TLDR is that we want to make sure at the time SetOnlySystemOatFiles is called, no oat file from /data have been loaded. The check currently gives that as the only oat files that used to be loaded were the boot classpath).

Patch-set: 2
Reviewer: Gerrit User 1038443 <1038443@85c56323-6fa9-3386-8a01-6480fb634889>