Update patch set 5
Patch Set 5:

(3 comments)

> Patch Set 4:
> 
> (3 comments)
> 
> I actually reviewed this change but forgot to post my comment.
> 
> I guess there's not real concerns in adding this support, but what are some use cases you think about? I view IMC as something purely dynamic, and now we're trying to capture it in a  static context.

Ah, sorry, how rude of me. I assumed you'd look at the relation chain of CLs.

The use case is pretty niche. Yes, IMC is purely dynamic, so this would never make any sense in  AOT. If you have a secondary dex which has IMC in its context, dex2oat should still fail because that dynamic dex is not available. That is also why I'm making it impossible to open dex files from an IMC spec.

What the CLs on top of this are doing, however, is building a small vdex created and loaded at runtime, which caches verification results. That vdex needs a class loader context and it is likely that context will contain an IMC. But at that point the IMC is available and if checksums match, the vdex data can be used. Does that make sense?

Patch-set: 5
Reviewer: Gerrit User 1059815 <1059815@85c56323-6fa9-3386-8a01-6480fb634889>