Partially disable building ART for host on Mac.
We don't use them, so no need to build them.

Running ART on Mac is not supported in the first place and
would require significant effort to make it work.

Building ART for the device/target on Mac still works.

The main motivation is remove custom elf.h from the code base.

Bug: 73344075
Test: "m -j40" on Mac.
Change-Id: I4b6f7aa7b291dcb6d751374ecf8d15f35877bd2e