Update patch set 10
Patch Set 10:

(1 comment)

> Patch Set 10:
> 
> > > > > > Can the pooling be done on the Java side? I don't see a reason for the ART complexity.
> > > > > 
> > > > > That's not a bad idea
> > > > 
> > > > On a second thought though... It also requires keeping a global state somewhere (ThreadPool in VMRuntime?) and a low-level API to wait (needed for tests), so I don't think it saves much complexity and moves the remaining complexity to libcore. Not a clear win.
> > > 
> > > There's a standard thread pool on the Java side (common ForkJoinPool). Also, I don't see why optimizing this is really necessary. I'd think we expect a small number of these, so maybe even a thread per dex file is OK?
> > 
> > Expect: yes. Do we want to put in a cap: also yes.
> > 
> > > 
> > > For tests, sleep.
> > 
> > I would expect you to be the champion of minimizing potential to flakiness. Honestly I don't see how we would save much by moving it to libcore. ART's ThreadPool is as simple as it gets.
> 
> Except how many pools do we have now?

I agree that there is a cost to maintaining the code but I still believe having a separate thread pool for this is the right thing to do because it gives us the most control over its configuration. I would be open to merging this with the JIT pool (including not running this when JIT is disabled) but that might make the code even weirder. WDYT?

Patch-set: 10
Reviewer: Gerrit User 1059815 <1059815@85c56323-6fa9-3386-8a01-6480fb634889>