Update patch set 16
Patch Set 16:

> Patch Set 15:
> 
> Looks like removing RegisterDexFile meant that the dex files are not getting the correct hiddenapi domain, i.e. FindClass and verifier do not register the dex files themselves.
> I'll have to investigate

It ended up being much simpler - 674-hiddenapi changes the hiddenapi config in Runtime frequently and there is no thread synchronization. The test became flaky because it saw old versions of `hiddenapi_exemptions_list_`. The proper solution would be to split the test into multiple smaller tests (there is a TODO for that), but for now I'm just switching it to PathClassLoader (the choice of CL doesn't matter for the test).

Patch-set: 16
Reviewer: Gerrit User 1059815 <1059815@85c56323-6fa9-3386-8a01-6480fb634889>