Exclude internal dependencies ASAN might add.
This required extending the superfluous files check to handle glob patterns.

This also corrects a bug in check_no_superfluous_files that caused it to not
check directories correctly for the host apex (due to trailing '/' in the
call to HostApexProvider.read_dir), so a few missing entries is added there.

Bug: 124293228
Test: art/build/apex/runtests.sh (with and without SANITIZE_TARGET/SANITIZE_HOST).
Change-Id: I83030b5b696fd80f42dd8722737b02fdf3a4089d