Support boot classpath checksum without image
Add special case to ImageSpace::GetBootClassPathChecksums to handle
case when image space is empty. In that case, we set the component
count and boot image checksum to zero and treat all dex files as the
boot classpath "tail", encoding their dex location checksums one by
one.

This fixes a DCHECK failure in 692-vdex-inmem-loader.

Test: ./art/test/testrunner/run_build_test_target.py \
      art-interpreter-no-image
Bug: 72131483

Change-Id: I3deb2c4704a010214c2f4287cc3010cb5fb450dd