Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Not convinced the "noncontraversial" dependencies are the things likely to cause problems here. Just grepping open, shows some system dependencies. Need some confidence that those are stable.

It uses socket_local_client to connect to a unix socket to talk to tombstoned. So we rely on the stability of that socket protocol. It looks like the socket connection is only used from [1] to send a free-form(?) printable stack trace/dump message. So it seems not too bad to plumb it through libartpalette. Shall we do that?

[1] http://cs/aosp-master/art/runtime/signal_catcher.cc?l=112-136&rcl=3b23f7568077a10cb3dfd0bb7d1c5f4c035d65dc

Patch-set: 2
Reviewer: Gerrit User 1373864 <1373864@85c56323-6fa9-3386-8a01-6480fb634889>