Update patch set 10
Patch Set 10:

Thanks for being active and addressing the feedback promptly!

I have a few general comments on the CL which I think should be assessed/discussed before proceeding with in-depth review.

1. Benchmarking
a) Code size and compile time impact must be evaluated; compiling boot*.oat (and maybe some big real-world applications) could be a good target.
b) In the benchmark results you've provided does "-" mean regressions? Have you analyzed the regressions?

2. Links and references
This is a classic optimization described and implemented in various places (including major native compilers). Could you provide links so it is easier to understand for people what algorithm you are implementing, what its complexity is, whether there are alternative solutions, etc.

3. Motivation
It seems that the range of loops to which the optimization is applied is quite small.
More generic question: what is the gain of the new optimization and pass in comparison to overhead (compile time, code size, code maintenance)?

Patch-set: 10
Reviewer: Gerrit User 1089808 <1089808@85c56323-6fa9-3386-8a01-6480fb634889>