Improve Mutex performance
Eliminate the separate seqentially consistent contention counter load
in Mutex::ExclusiveUnlock by putting the contenion counter in the
Mutex's state word.

Replace some CHECK_GE checks with CHECK_GT. We were checking quantities
intended to be non-negative against >= 0 just before decrementing them.

Remove a pointless volatile declaration.

This does not yet address the ReaderWriterMutex issue, which can't
easily be handled the same way.

This is intended as a short-term fix for a recent performance regression:
We should try to reduce the number of mutex implementations. If we have
our own in ART, we should almost certainly make a type distinction between
recursive and non-recursive mutexes to avoid the recursion count
manipulation where it's not needed.

Bug: 111835365
Test: Build and boot AOSP. TreeHugger.
Change-Id: Ie53c65f2ce774a8cb4d224e2c1b3a110eb880f0c