Update patch set 3
Patch Set 3: Presubmit-Ready+1

This updates the 005-annotations test to pass the --intermediate flag when compiling with D8. The D8 compiler is moving towards removing all annotations with a class-retention policy for non-intermediate builds and this is needed for the test to still pass after that change. The failure without this falg, would be caused by the test setting target sdk level to M and then inspecting the annotation and expecting it to be visible/present.

Patch-set: 3
Reviewer: Gerrit User 1038443 <1038443@85c56323-6fa9-3386-8a01-6480fb634889>
CC: Gerrit User 1121678 <1121678@85c56323-6fa9-3386-8a01-6480fb634889>
Reviewer: Gerrit User 1241761 <1241761@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=+1