[WIP] StringBuilder append pattern for float/double.
Results for added benchmarks on taimen:
  timeAppendStringAndDouble: ~450ns -> 420ns
  timeAppendStringAndFloat: ~450ns -> 420ns
  timeAppendStringAndHugeDouble: ~1650ns -> ~1750ns
  timeAppendStringAndHugeFloat: ~1110ns -> ~1130ns
  timeAppendStringDoubleStringAndFloat: ~660ns -> ~780ns
TODO: The Huge D/F regressions are difficult to explain, in
addition to the non-Huge D/F, they execute some extra code
identical for "before" and "after". Maybe it's because of
some cache effects. These are also the noisiest benchmarks.
The AppendSDSF regression is probably because of the slow
transitions between managed and native code (which we do for
two FP arguments) outweigh avoiding allocations. Note that
these benchmarks test the best-case scenario for "before" as
the StringBuilder's internal buffer is not reallocated.

Boot image size changes are insignificant (few matches).

Test: Added tests to 697-checker-string-append
Test: m test-art-host-gtest
Test: testrunner.py --host --optimizing
Test: aosp_taimen-userdebug boots.
Test: testrunner.py --target --optimizing
Bug: 19575890
Change-Id: I9cf38c2d615a0a2b14255d18588a694d8870aae5